requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
pypdf>=4.2.0
gpt4all>=2.8.2
fastapi
uvicorn
pywebview
python-multipart
